requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
langchain
langchain-openai
langchain-community
psycopg2-binary
langchain-supabase
tiktoken
